+++ sys/pami.h
@@ -1207,7 +1207,7 @@ extern "C"
   pami_result_t PAMI_Geometry_algorithms_num (pami_context_t context,
                                             pami_geometry_t geometry,
                                             pami_xfer_type_t coll_type,
-                                            int              *lists_lengths);
+                                            size_t              *lists_lengths);
 
   /**
    * \brief fills in the protocols and attributes for a set of algorithms
@@ -1233,10 +1233,10 @@ extern "C"
                                              pami_xfer_type_t  colltype,
                                              pami_algorithm_t *algs0,
                                              pami_metadata_t  *mdata0,
-                                             int              num0,
+                                             size_t              num0,
                                              pami_algorithm_t *algs1,
                                              pami_metadata_t  *mdata1,
-                                             int              num1);
+                                             size_t              num1);
 
   /**
    * \brief Free any memory allocated inside of a geometry.
